+++ crates/rich/src/lib.rs
@@ -26,6 +26,7 @@ pub mod columns;
 pub mod console;
 pub mod constrain;
 pub mod emoji;
+mod emoji_codes;
 pub mod errors;
 pub mod filesize;
 pub mod json;